Add support for 'or' query operator
Now query is built recursively.

Using 'or' allows combine subqueries with sql OR operator. As query is built recursively, OR operands can be lists of subquires (that are concatenated using AND), that can have OR operands inside them and so on.

So every possible query structure can be created with addition of OR operand